==> 8)Eli5 Result and comparison.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

np.random.seed(1)

from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score
from sklearn.model_selection import cross_val_score
import eli5
from eli5.sklearn import PermutationImportance
from sklearn.linear_model import LogisticRegression

RAW_DATA_FILEPATH = '/users/eric/PycharmProjects/testico/api_data.json'
URL_FILEPATH = '/users/eric/PycharmProjects/testico/ico_urls.txt'
TOKEN_FILE = '/users/eric/PycharmProjects/testico/token.csv'
FULLDATA_FILE = '/users/eric/PycharmProjects/testico/fulldata.csv'
NUMERICAL_DATA_FILE = '/users/eric/PycharmProjects/testico/full_num.csv'
TEXT_DATA_FILE = '/users/eric/PycharmProjects/testico/textfeature_dataframe.csv'

full_num = pd.read_csv(NUMERICAL_DATA_FILE, index_col=0)


# Ignore last label: ROI
feature_names = full_num.columns.tolist()[:-1]
#import ipdb; ipdb.set_trace()
y = full_num['binary_ROI']
X = full_num[feature_names]
train_X, val_X, train_y, val_y = train_test_split(X, y, random_state=1)
clf = RandomForestClassifier(random_state=0, class_weight="balanced")
my_model = clf.fit(train_X, train_y)

perm1 = PermutationImportance(my_model, random_state=1).fit(val_X, val_y)
eli5.show_weights(perm1, feature_names=val_X.columns.tolist())



X_git


# binarize_Git
feature_names2 = full_num.columns.tolist()[:-2]
X2 = full_num[feature_names2]



import warnings
warnings.filterwarnings("ignore", category=FutureWarning)

TOKEN_GIT_FEATURE_FILE = '/users/eric/PycharmProjects/testico/token_with_github_feature.csv'
FULLDATA_PLUS_GIT = '/users/eric/PycharmProjects/testico/fulldata_plus_git.csv'
NUMERICAL_DATA_PLUS_GIT_FILE = '/users/eric/PycharmProjects/testico/num_plus_git.csv'

full_num_git = pd.read_csv(NUMERICAL_DATA_PLUS_GIT_FILE, index_col=0)




# Ignore last label: ROI
feature_names_git = full_num_git.columns.tolist()[:-1]
#import ipdb; ipdb.set_trace()
y_git = full_num_git['binary_ROI']
X_git = full_num_git[feature_names_git]
train_X_git, val_X_git, train_y_git, val_y_git = train_test_split(X_git, y_git, random_state=1)
my_model = RandomForestClassifier(random_state=0).fit(train_X_git, train_y_git)
clf = RandomForestClassifier(random_state=0, class_weight="balanced")
# perm2 = PermutationImportance(my_model, random_state=1).fit(val_X_git, val_y_git)
# eli5.show_weights(perm2, feature_names_git=val_X_git.columns.tolist())


# y_pred = my_model.predict(val_X)
# accuracy = accuracy_score(val_y, y_pred)
# print(accuracy)



# train, test = train_test_split(full_num, test_size=0.33)
# array_train = train.values
# array_test = test.values

# x_train = array_train[:, 0:-1]
# y_train = array_train[:, -1]

# x_test = array_test[:, 0:-1]
# y_test = array_test[:, -1]

# model = XGBClassifier()
# model.fit(x_train, y_train)
# #
# # y_pred = model.predict(x_test)
# # accuracy = accuracy_score(y_test, y_pred)
# # print(accuracy)

# logmodel = LogisticRegression()
# # logmodel.fit(train_X,train_y )


# # cross-validation with 10folds
# RFCscores = cross_val_score(my_model, X, y, cv=10)
# print(np.mean(RFCscores))

# fullvalue = full_num .values
# traindata = fullvalue[:, 0:-1]
# testdata = fullvalue[:, -1]
# xgbc = model.fit(x_train, y_train)
# XGBscores = cross_val_score(xgbc, traindata, testdata, cv=10)
# print(np.mean(XGBscores))

# LRscores = cross_val_score(logmodel, X_git, y_git, cv=10)
# print(np.mean(LRscores))

from sklearn.model_selection import LeaveOneOut
loo = LeaveOneOut()
X_np = np.array(X)
y_np = np.array(y)
loo.get_n_splits(X_np)
# loo.get_n_splits(X_git)
X_np_git = np.array(X_git)
y_np_git = np.array(y_git)
loo.get_n_splits(X_np_git)
X_np2 = np.array(X2)
loo.get_n_splits(X_np2)

# scores = []
# for train_index, test_index in loo.split(X_np):
#     X_train_loo, X_test_loo = X_np[train_index], X_np[test_index]
#     y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
#     clf2_loo = clf.fit(X_train_loo, y_train_loo)
#     y_pred_loo = clf2.predict(X_test_loo)
#     score = accuracy_score(y_test_loo, y_pred_loo)
#     scores.append(score)
# average_score = sum(scores) / float(len(scores))
# average_score

    
    

# y_pred = my_model.predict(val_X)
# accuracy = accuracy_score(val_y, y_pred)
# print(accuracy)



# train, test = train_test_split(full_num, test_size=0.33)
# array_train = train.values
# array_test = test.values

# x_train = array_train[:, 0:-1]
# y_train = array_train[:, -1]

# x_test = array_test[:, 0:-1]
# y_test = array_test[:, -1]

 #(scale_pos_weight=2.6486) #很重要的参数 提升了结果
# model.fit(x_train, y_train)

# y_pred = model.predict(x_test)
# accuracy = accuracy_score(y_test, y_pred)
# print(accuracy)




clf = RandomForestClassifier(random_state=0, class_weight='balanced')
scores_clf = cross_val_score(clf, X, y, cv=10)
scores_clf2 = cross_val_score(clf, X2,y, cv=10)

clf = RandomForestClassifier(random_state=0, class_weight='balanced')
scores_clf_git = cross_val_score(clf ,X_git ,y_git , cv =10) 
print(np.mean(scores_clf2))
print(np.mean(scores_clf))

print(np.mean(scores_clf_git))



scores_clf2_loo = []
for train_index, test_index in loo.split(X_np2):
    X_train_loo, X_test_loo = X_np2[train_index], X_np2[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    clf_loo = clf.fit(X_train_loo, y_train_loo)
    y_pred_loo = clf_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_clf2_loo.append(score)
average_score_clf2 = sum(scores_clf2_loo) / float(len(scores_clf2_loo))
average_score_clf2

scores_clf_loo = []
for train_index, test_index in loo.split(X_np):
    X_train_loo, X_test_loo = X_np[train_index], X_np[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    clf_loo = clf.fit(X_train_loo, y_train_loo)
    y_pred_loo = clf_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_clf_loo.append(score)
average_score_clf = sum(scores_clf_loo) / float(len(scores_clf_loo))
average_score_clf


scores_clf_git_loo = []
for train_index, test_index in loo.split(X_np_git):
    X_train_loo, X_test_loo = X_np_git[train_index], X_np_git[test_index]
    y_train_loo, y_test_loo = y_np_git[train_index], y_np_git[test_index]
    clf_loo = clf.fit(X_train_loo, y_train_loo)
    y_pred_loo = clf_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_clf_git_loo.append(score)
average_score_clf_git = sum(scores_clf_git_loo) / float(len(scores_clf_git_loo))
average_score_clf_git






logmodel = LogisticRegression(random_state=0, class_weight='balanced')
scores_log = cross_val_score(logmodel, X, y, cv=10)
scores_log2 = cross_val_score(logmodel, X2, y, cv=10)

scores_log_git = cross_val_score(logmodel, X_git, y_git, cv=10)
print(np.mean(scores_log2))

print(np.mean(scores_log))

print(np.mean(scores_log_git))

scores_log2_loo = []
for train_index, test_index in loo.split(X_np2):
    X_train_loo, X_test_loo = X_np2[train_index], X_np2[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    log_loo = logmodel.fit(X_train_loo, y_train_loo)
    y_pred_loo = log_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_log2_loo.append(score)
average_score_log2 = sum(scores_log2_loo) / float(len(scores_log2_loo))
average_score_log2

scores_log_loo = []
for train_index, test_index in loo.split(X_np):
    X_train_loo, X_test_loo = X_np[train_index], X_np[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    log_loo = logmodel.fit(X_train_loo, y_train_loo)
    y_pred_loo = log_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_log_loo.append(score)
average_score_log = sum(scores_log_loo) / float(len(scores_log_loo))
average_score_log



scores_log_git_loo = []
for train_index, test_index in loo.split(X_np_git):
    X_train_loo, X_test_loo = X_np_git[train_index], X_np_git[test_index]
    y_train_loo, y_test_loo = y_np_git[train_index], y_np_git[test_index]
    log_loo = logmodel.fit(X_train_loo, y_train_loo)
    y_pred_loo = log_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_log_git_loo.append(score)
average_score_log_git = sum(scores_log_git_loo) / float(len(scores_log_git_loo))
average_score_log_git



# fullvalue = full_num .values
# traindata = fullvalue[:, 0:-1]
# testdata = fullvalue[:, -1]
# model = XGBClassifier()

# #xgbc = model.fit(train_X, train_y)
# scores = cross_val_score(model, X, y, cv=10)
# print(scores)
# print(np.mean(scores))
# print(np.var(scores))

xgb = XGBClassifier(scale_pos_weight=2.6486) #很重要的参数 提升了结果
scores2 = cross_val_score(xgb,X2,y, cv=10)
scores = cross_val_score(xgb, X, y, cv=10)
scores_git = cross_val_score(xgb, X_git, y_git, cv=10)
#scores_loo = cross_val_score(model2, X_git ,y_git, cv = X.shape[0]) 
print(np.mean(scores2))
print(np.mean(scores))
print(np.mean(scores_git))
#print(np.means(scores_loo))

xgbperm2 = XGBClassifier(train_X_git)
perm2 = PermutationImportance(xgb, random_state=1).fit(val_X, val_y)
eli5.show_weights(perm2, feature_names=val_X.columns.tolist())



scores_xgb2 = []
for train_index, test_index in loo.split(X_np2):
    X_train_loo, X_test_loo = X_np2[train_index], X_np2[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    xgb_loo = xgb.fit(X_train_loo, y_train_loo)
    y_pred_loo = xgb_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_xgb2.append(score)
average_score_xgb2 = sum(scores_xgb2) / float(len(scores_xgb2))
average_score_xgb2

scores_xgb = []
for train_index, test_index in loo.split(X_np):
    X_train_loo, X_test_loo = X_np[train_index], X_np[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    xgb_loo = xgb.fit(X_train_loo, y_train_loo)
    y_pred_loo = xgb_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_xgb.append(score)
average_score_xgb = sum(scores_xgb) / float(len(scores_xgb))
average_score_xgb

scores_xgb_git = []
for train_index, test_index in loo.split(X_np_git):
    X_train_loo, X_test_loo = X_np_git[train_index], X_np_git[test_index]
    y_train_loo, y_test_loo = y_np_git[train_index], y_np_git[test_index]
    xgb_loo = xgb.fit(X_train_loo, y_train_loo)
    y_pred_loo = xgb_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_xgb_git.append(score)
average_score_xgb_git = sum(scores_xgb_git) / float(len(scores_xgb_git))
average_score_xgb_git

perm1 = PermutationImportance(my_model, random_state=1).fit(val_X, val_y)
eli5.show_weights(perm1, feature_names=val_X.columns.tolist())
perm2 =


from sklearn import svm
svm_model = svm.SVC(class_weight='balanced')
scores_svm = cross_val_score(svm_model, X, y, cv=10)
scores_svm2 = cross_val_score(svm_model, X2,y, cv =10)
scores_svm_git = cross_val_score(svm_model, X_git, y_git, cv=10)

print(np.mean(scores_svm2))
print(np.mean(scores_svm))


print(np.mean(scores_svm_git))


scores_svm2_loo = []
for train_index, test_index in loo.split(X_np2):
    X_train_loo, X_test_loo = X_np2[train_index], X_np2[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    svm_loo = svm_model.fit(X_train_loo, y_train_loo)
    y_pred_loo = svm_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_svm2_loo.append(score)
average_score_log = sum(scores_svm2_loo) / float(len(scores_svm2_loo))
average_score_log

scores_svm_loo = []
for train_index, test_index in loo.split(X_np):
    X_train_loo, X_test_loo = X_np[train_index], X_np[test_index]
    y_train_loo, y_test_loo = y_np[train_index], y_np[test_index]
    svm_loo = svm_model.fit(X_train_loo, y_train_loo)
    y_pred_loo = svm_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_svm_loo.append(score)
average_score_log = sum(scores_svm_loo) / float(len(scores_svm_loo))
average_score_log

scores_svm_git_loo = []
for train_index, test_index in loo.split(X_np_git):
    X_train_loo, X_test_loo = X_np_git[train_index], X_np_git[test_index]
    y_train_loo, y_test_loo = y_np_git[train_index], y_np_git[test_index]
    svm_loo = svm_model.fit(X_train_loo, y_train_loo)
    y_pred_loo = svm_loo.predict(X_test_loo)
    score = accuracy_score(y_test_loo, y_pred_loo)
    scores_svm_git_loo.append(score)
average_score_xgb_git = sum(scores_svm_git_loo) / float(len(scores_svm_git_loo))
average_score_xgb_git

model3 = XGBClassifier(random_state=0).fit(train_X, train_y)
perm2 = PermutationImportance(model3, random_state=0).fit(val_X, val_y)
eli5.show_weights(perm2, feature_names=val_X.columns.tolist())

X

X.shape


y.sum()

y.groupyby(by=axis1).size()

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

np.random.seed(1)

from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score
from sklearn.model_selection import cross_val_score
import eli5
from eli5.sklearn import PermutationImportance

RAW_DATA_FILEPATH = '/users/eric/PycharmProjects/testico/api_data.json'
URL_FILEPATH = '/users/eric/PycharmProjects/testico/ico_urls.txt'
TOKEN_FILE = '/users/eric/PycharmProjects/testico/token.csv'
FULLDATA_FILE = '/users/eric/PycharmProjects/testico/fulldata.csv'
NUMERICAL_DATA_FILE = '/users/eric/PycharmProjects/testico/full_num.csv'
TEXT_DATA_FILE = '/users/eric/PycharmProjects/testico/textfeature_dataframe.csv'
TOKEN_GIT_FEATURE_FILE = '/users/eric/PycharmProjects/testico/token_with_github_feature.csv'
FULLDATA_PLUS_GIT = '/users/eric/PycharmProjects/testico/fulldata_plus_git.csv'
NUMERICAL_DATA_PLUS_GIT_FILE = '/users/eric/PycharmProjects/testico/num_plus_git.csv'

full_num_git = pd.read_csv(NUMERICAL_DATA_PLUS_GIT_FILE, index_col=0)

# Ignore last label: ROI
feature_names_new = full_num_git.columns.tolist()[:-1]
#import ipdb; ipdb.set_trace()
y = full_num_git['binary_ROI']
X = full_num_git[feature_names_new]
train_X, val_X, train_y, val_y = train_test_split(X, y, random_state=1)
clf = RandomForestClassifier(random_state=0, class_weight="balanced")
#my_model = clf.fit(train_X, train_y)

y_pred = my_model.predict(val_X)
accuracy = accuracy_score(val_y, y_pred)
print(accuracy)

scores_RFC = cross_val_score(clf, X, y, cv=10)
print(np.mean(scores_RFC))

from sklearn.linear_model import LogisticRegression
logmodel = LogisticRegression()
logmodel.fit(train_X,train_y )
scores_LR = cross_val_score(logmodel, X, y, cv=10)
print(np.mean(scores_LR))

FULLDATA_PLUS_GIT = '/users/eric/PycharmProjects/testico/fulldata_plus_git.csv'

fulldata = pd.read_csv(FULLDATA_PLUS_GIT)


fulldata.shape

gitnum = fulldata.dropna(subset=['Github URL'], inplace=True)

fulldata.shape

fulldata2 = pd.read_csv(FULLDATA_PLUS_GIT)

fulldata2.shape

fulldata2.dropna(subset=['ROI'], inplace=True)

fulldata2.shape

fulldata2.dropna(subset=['Github URL'], inplace=True)

fulldata2.shape

fulldata3 = pd.read_csv(FULLDATA_PLUS_GIT)

#  'watchers','stars','forks','branches', 'commits'
fulldata3['branches'].mean()
